+++ docs/src/markdown/about/changelog.md
@@ -7,6 +7,7 @@ icon: lucide/scroll-text
 
 -   **NEW**: Implement special gamuts (`visible-spectrum`, `pointer-gamut`, etc.) as `Gamut` plugins allowing the
     expansion of _special_ gamuts in the future.
+-   **NEW**: Minor tweaks to `macadam-limits` calculations.
 
 ## 8.12.1